requirements.txt
gspread
gspread-dataframe
google-auth
pandas
requests


